samay-v3/requirements.txt

seleniumbase==4.40.6
undetected-chromedriver==3.5.5

imap-tools==1.11.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.148.0
beautifulsoup4==4.13.4

python-dotenv==1.0.1
redis==5.1.1
requests==2.31.0

# Web API dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0

asyncio-mqtt==0.16.2
cryptography==42.0.8